requirements.txt
django
python-dotenv
requests
gunicorn
sarvamai 